+++ service/internal/cloudconfig/template/etcd3_attach_dep_service.go
@@ -27,8 +27,9 @@ ExecStart=/bin/bash -c "docker run --rm -i \
       --volume-device-label=etcd \
       --volume-tag-key=Name \
       --volume-tag-value={{ .MasterEtcdVolumeName }}"
-ExecStartPost=/usr/bin/systemctl daemon-reload
-ExecStartPost=/usr/bin/systemctl restart systemd-networkd
+# use 'networkctl reload' instead of restarting systemd-networkd to avoid bug in systemd 
+# https://github.com/systemd/systemd/issues/18108
+ExecStartPost=/usr/bin/networkctl reload
 
 [Install]
 WantedBy=multi-user.target