+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fix issues with etcd initial cluster resolving into ELB and causing errors.
+- Update `k8scloudconfig` to version `v10.5.0` to support kubernetes `v1.20`.
+- Use `networkctl reload` for managing networking to avoid bug in `systemd`.
 
 ## [10.2.0] - 2021-02-08